feat(MovieListFragment): 기존 액티비티 뷰로직 이동
