+++ app/src/main/java/woowacourse/movie/data/AdData.kt
@@ -1,7 +1,7 @@
 package woowacourse.movie.data
 
 import woowacourse.movie.R
-import woowacourse.movie.presentation.main.MovieItem
+import woowacourse.movie.presentation.movielist.MovieItem
 
 object AdData {
     val ads = List(10000) { MovieItem.Ad(R.drawable.woowa_ad) }
