+++ app/src/main/res/layout/fragment_movie_list.xml
@@ -1,18 +1,17 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.main.MainActivity">
+    tools:context=".presentation.movielist.MovieListFragment">
 
-
-    <androidx.fragment.app.FragmentContainerView
+    <androidx.recyclerview.widget.RecyclerView
+        android:id="@+id/recyclerMainMovie"
         android:layout_width="match_parent"
         android:layout_height="0dp"
-        android:id="@+id/main_fragment_container"
-        tools:layout="@layout/fragment_movie_list"
+        app:layoutManager="LinearLayoutManager"
+        app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintBottom_toBottomOf="parent" />
-
+        tools:listitem="@layout/movie_list_item" />
 </androidx.constraintlayout.widget.ConstraintLayout>
