+++ app/src/main/res/values/strings.xml
@@ -17,5 +17,7 @@
     <string name="dialog_choice_confirm_message">정말 예매하시겠습니까?</string>
     <string name="dialog_choice_positive_button">예매 완료</string>
     <string name="dialog_choice_negative_button">취소</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 
 </resources>
